🔨 Refactor: common Entity Package 수정
