+++ src/main/java/com/cheeup/domain/common/Major.java
@@ -1,4 +1,4 @@
-package com.cheeup.domain;
+package com.cheeup.domain.common;
 
 import jakarta.persistence.*;
 import lombok.*;
@@ -8,12 +8,16 @@ import lombok.*;
 @Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class Skills {
+@Table(name = "majors")
+public class Major {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @Column(nullable = false, length = 50)
     private String name;
-}
 
+    @Column(nullable = false)
+    private Integer category;
+}
